smartcontract-analysis/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
python-multipart